Added a visited urls "cache" that is just a map[string]bool.
NOTE: The map[string]bool is synced with a channel. This means that some
readers of the definite cache might not be completely up to date. To fix
this feature, the writes to the map should have a RW Mutex but I don't
want to use that since it will keep everyone waiting and will be slow as
hell. Instead, the readers are going to be a bit "behind" on the actual
visited urls but that's fine for my usecase.

Also printing of visited sites is now optional since I added an explicit
flag to allow the prints. I like to use this but someone might not like
the prints.