Once upon a time
Once upon a time I thought I would separate the href grabber and HTML
parser so I made it so the start of the goroutine would add the href
link back so it could be consumed by the HTML parser. This ended up
being an overly complex solution so I just made it so they're all done
in the same loop. I'm getting rid of the thing that adds the href link
back to the queue after consuming it.