In-memory cache for visited urls added. (~82% accurate)
This can be disabled by setting visitDuplicates to true. There exists a
setter for this.
To my calculations the visitedUrls mutex map works 82% of the times.
There is a bug/feature that let's a url to be revisited 17.9% of the
time. I'll fix this later if I can but I think this is a result of
something to do with multiple goroutines doing stuff. I guess it's good
enough for now since I got a bunch of other shit to do.